+++ Sources/TMDB/Coordinator/Coordinator.swift
@@ -1,81 +1,28 @@
 import Combine
 import SwiftUI
 
-public enum TabRoute: Hashable {
-    case nowPlaying
-    case upcoming
-    case profile
-
-    public var title: String {
-        switch self {
-        case .nowPlaying: return "Now Playing"
-        case .upcoming: return "Upcoming"
-        case .profile: return "Profile"
-        }
-    }
-
-    public var iconName: String {
-        switch self {
-        case .nowPlaying: return "play.circle"
-        case .upcoming: return "magnifyingglass"
-        case .profile: return "list.bullet"
-        }
-    }
-}
-
-// MARK: - NavigationState.swift
-
-@available(iOS 16.0, *)
-public class NavigationState: ObservableObject {
-    @Published public var path = NavigationPath()
-    @Published public var sheet: (any Route)?
-
-    public func navigate(to route: TMDBRoute) {
-        path.append(route)
-        objectWillChange.send() // Explicitly notify about the change
-    }
-
-    public func push(_ route: any Route) {
-        path.append(route)
-    }
-
-    public func present(_ route: any Route) {
-        sheet = route
-    }
-
-    public func pop() {
-        path.removeLast()
-    }
-
-    public func popToRoot() {
-        while !path.isEmpty {
-            path.removeLast()
-        }
-    }
-
-    public func dismiss() {
-        sheet = nil
-    }
-}
-
 // MARK: - Coordinator.swift
-
 @available(iOS 16.0, *)
 @MainActor
 public class Coordinator: ObservableObject {
     @Published public var selectedTab: TabRoute
     @Published private(set) var navigationStates: [TabRoute: NavigationState]
+    @Published private(set) var tabBarHiddenStates: [TabRoute: Bool] // Tracks tab bar visibility
     public let tabList: [TabRoute]
+
     public init(tabList: [TabRoute]) {
         self.tabList = tabList
         selectedTab = tabList[0]
 
-        // Initialize navigation state for each tab
+        // Initialize navigation and visibility states
         var states: [TabRoute: NavigationState] = [:]
+        var hiddenStates: [TabRoute: Bool] = [:]
         for tab in tabList {
             states[tab] = NavigationState()
+            hiddenStates[tab] = false // Tab bar visible by default
         }
         navigationStates = states
+        tabBarHiddenStates = hiddenStates
     }
 
     public func path(for tab: TabRoute) -> Binding<NavigationPath> {
@@ -85,6 +32,10 @@ public class Coordinator: ObservableObject {
             },
             set: { [weak self] newValue in
                 self?.navigationStates[tab]?.path = newValue
+                // Update visibility based on navigation depth
+                if let count = self?.navigationStates[tab]?.path.count {
+                    self?.tabBarHiddenStates[tab] = count > 0
+                }
             }
         )
     }
@@ -105,7 +56,25 @@ public class Coordinator: ObservableObject {
     public func navigate(to route: TMDBRoute, in tab: TabRoute) {
         if let state = navigationStates[tab] {
             state.navigate(to: route)
-            objectWillChange.send() // Explicitly notify about the change
+            tabBarHiddenStates[tab] = true // Hide tab bar on navigation
+            objectWillChange.send()
+        }
+    }
+
+    public func pop(in tab: TabRoute) {
+        if let state = navigationStates[tab] {
+            state.pop()
+            // Show tab bar if back at root
+            tabBarHiddenStates[tab] = state.path.isEmpty
+            objectWillChange.send()
+        }
+    }
+
+    public func popToRoot(in tab: TabRoute) {
+        if let state = navigationStates[tab] {
+            state.popToRoot()
+            tabBarHiddenStates[tab] = false // Show tab bar at root
+            objectWillChange.send()
         }
     }
 }
